+++ spring-cloud-starter-tencent-polaris-router/src/main/java/com/tencent/cloud/polaris/router/beanprocessor/LoadBalancerInterceptorBeanPostProcessor.java
@@ -17,13 +17,7 @@
 
 package com.tencent.cloud.polaris.router.beanprocessor;
 
-import java.util.List;
-
-import com.tencent.cloud.common.metadata.StaticMetadataManager;
-import com.tencent.cloud.common.util.BeanFactoryUtils;
-import com.tencent.cloud.polaris.router.RouterRuleLabelResolver;
 import com.tencent.cloud.polaris.router.resttemplate.PolarisLoadBalancerInterceptor;
-import com.tencent.cloud.polaris.router.spi.SpringWebRouterLabelResolver;
 
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.BeanFactory;
@@ -53,12 +47,8 @@ public class LoadBalancerInterceptorBeanPostProcessor implements BeanPostProcess
 		if (bean instanceof LoadBalancerInterceptor) {
 			LoadBalancerRequestFactory requestFactory = this.factory.getBean(LoadBalancerRequestFactory.class);
 			LoadBalancerClient loadBalancerClient = this.factory.getBean(LoadBalancerClient.class);
-			List<SpringWebRouterLabelResolver> routerLabelResolvers = BeanFactoryUtils.getBeans(factory, SpringWebRouterLabelResolver.class);
-			StaticMetadataManager staticMetadataManager = this.factory.getBean(StaticMetadataManager.class);
-			RouterRuleLabelResolver routerRuleLabelResolver = this.factory.getBean(RouterRuleLabelResolver.class);
 
-			return new PolarisLoadBalancerInterceptor(loadBalancerClient, requestFactory,
-					routerLabelResolvers, staticMetadataManager, routerRuleLabelResolver);
+			return new PolarisLoadBalancerInterceptor(loadBalancerClient, requestFactory);
 		}
 		return bean;
 	}
